refactor player style
